Move domain informer to virt-handler
